src/TestRunner,Printer: Prevent accidental inclusion of HardwareSerial
Remove reference to `SERIAL_PORT_MONITOR` from `Printer.cpp`. Add it to the
`setupRunner()` method in the `TestRunner.h` header file. Move any
implementation code that references `setupRunner()` (i.e. `listTests()` and
`runTest()`) from `TestRunner.cpp` to `TestRunner.h`.

Saves about 800 bytes of flash on AVR programs which do not use the Serial port
at all. Saves about 3kB on ESP8266 and 5kB on ESP32, but the flash savings are
not as important on these processors because of their high flash memory
capacity.